+++ src/Vector/DBC/Signal.cpp
@@ -231,7 +231,7 @@ void Signal::encode(std::vector<uint8_t> & data, uint64_t rawValue)
     }
 }
 
-std::ostream & operator<<(std::ostream & os, Signal & signal)
+std::ostream & operator<<(std::ostream & os, const Signal & signal)
 {
     /* Name */
     os << " SG_ " << signal.name << ' ';