+++ src/Vector/DBC/Signal.h
@@ -190,7 +190,7 @@ public:
     void encode(std::vector<uint8_t> & data, uint64_t rawValue);
 };
 
-std::ostream & operator<<(std::ostream & os, Signal & signal);
+std::ostream & operator<<(std::ostream & os, const Signal & signal);
 
 }
 }