+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- operator<< argument is now const
+
 ## [2.0.2] - 2019-03-08
 ### Fixed
 - Support of newest Bison versions